modularized auth, vds and general utils to be shared across tools
