+++ community/langchain_community/utilities/tableau/simple_datasource_qa.py
@@ -1,40 +1,10 @@
-import os
 import json
 import re
-from typing import Dict, Any, List, Optional
-import requests
-import jwt
+from typing import Dict, Optional
 import logging
-from datetime import datetime, timedelta, timezone
-from uuid import uuid4
-from dotenv import load_dotenv
 
-
-def query_vds(api_key: str, datasource_luid: str, url: str, query: Dict[str, Any]) -> Dict[str, Any]:
-    full_url = f"{url}/api/v1/vizql-data-service/query-datasource"
-
-    payload = {
-        "datasource": {
-            "datasourceLuid": datasource_luid
-        },
-        "query": query
-    }
-
-    headers = {
-        'X-Tableau-Auth': api_key,
-        'Content-Type': 'application/json'
-    }
-
-    response = requests.post(full_url, headers=headers, json=payload)
-
-    if response.status_code == 200:
-        return response.json()
-    else:
-        error_message = (
-            f"Failed to query data source via Tableau VizQL Data Service. "
-            f"Status code: {response.status_code}. Response: {response.text}"
-        )
-        raise RuntimeError(error_message)
+from community.langchain_community.utilities.tableau.vizql_data_service import query_vds, query_vds_metadata
+from community.langchain_community.utilities.tableau.utils import json_to_markdown_table
 
 
 def get_headlessbi_data(payload: str, url: str, api_key: str, datasource_luid: str):
@@ -51,7 +21,7 @@ def get_headlessbi_data(payload: str, url: str, api_key: str, datasource_luid: s
         if not headlessbi_data or 'data' not in headlessbi_data:
             raise ValueError("Invalid or empty response from query_vds")
 
-        markdown_table = json_to_markdown(headlessbi_data['data'])
+        markdown_table = json_to_markdown_table(headlessbi_data['data'])
         return markdown_table
 
     except ValueError as ve:
@@ -83,51 +53,6 @@ def get_payload(output):
             raise ValueError("Invalid JSON format in the payload")
     else:
         raise ValueError("No JSON payload found in the parsed output")
-
-
-def json_to_markdown(json_data):
-    if isinstance(json_data, str):
-        json_data = json.loads(json_data)
-    # Check if the JSON data is a list and not empty
-    if not isinstance(json_data, list) or not json_data:
-        raise ValueError(f"Invalid JSON data, you may have an error or if the array is empty then it was not possible to resolve the query your wrote: {json_data}")
-
-    headers = json_data[0].keys()
-
-    markdown_table = "| " + " | ".join(headers) + " |\n"
-    markdown_table += "| " + " | ".join(['---'] * len(headers)) + " |\n"
-
-    for entry in json_data:
-        row = "| " + " | ".join(str(entry[header]) for header in headers) + " |"
-        markdown_table += row + "\n"
-
-    return markdown_table
-
-
-def query_vds_metadata(api_key: str, datasource_luid: str, url: str) -> Dict[str, Any]:
-    full_url = f"{url}/api/v1/vizql-data-service/read-metadata"
-
-    payload = {
-        "datasource": {
-            "datasourceLuid": datasource_luid
-        }
-    }
-
-    headers = {
-        'X-Tableau-Auth': api_key,
-        'Content-Type': 'application/json'
-    }
-
-    response = requests.post(full_url, headers=headers, json=payload)
-
-    if response.status_code == 200:
-        return response.json()
-    else:
-        error_message = (
-            f"Failed to obtain data source metadata from VizQL Data Service. "
-            f"Status code: {response.status_code}. Response: {response.text}"
-        )
-        raise RuntimeError(error_message)
 
 
 def get_values(api_key: str, url: str, datasource_luid: str, caption: str):
@@ -169,85 +94,6 @@ def augment_datasource_metadata(
     return json.dumps(prompt)
 
 
-def authenticate_tableau_user(
-        tableau_domain: str,
-        tableau_site: str,
-        tableau_api: str,
-        tableau_user: str,
-        jwt_client_id: str,
-        jwt_secret_id: str,
-        jwt_secret: str,
-        scopes: List[str],
-) -> Dict[str, Any]:
-    """
-    Authenticates a user to Tableau using JSON Web Token (JWT) authentication.
-
-    This function generates a JWT based on the provided credentials and uses it to authenticate
-    a user with the Tableau Server or Tableau Online. The JWT is created with a specified expiration
-    time and scopes, allowing for secure access to Tableau resources.
-
-    Args:
-        tableau_domain (str): The domain of the Tableau Server or Tableau Online instance.
-        tableau_site (str): The content URL of the specific Tableau site to authenticate against.
-        tableau_api (str): The version of the Tableau API to use for authentication.
-        tableau_user (str): The username of the Tableau user to authenticate.
-        jwt_client_id (str): The client ID used for generating the JWT.
-        jwt_secret_id (str): The key ID associated with the JWT secret.
-        jwt_secret (str): The secret key used to sign the JWT.
-        scopes (List[str]): A list of scopes that define the permissions granted by the JWT.
-
-    Returns:
-        Dict[str, Any]: A dictionary containing the response from the Tableau authentication endpoint,
-        typically including an API key or session that is valid for 2 hours and user information.
-    """
-    # Encode the payload and secret key to generate the JWT
-    token = jwt.encode(
-        {
-        "iss": jwt_client_id,
-        "exp": datetime.now(timezone.utc) + timedelta(minutes=5),
-        "jti": str(uuid4()),
-        "aud": "tableau",
-        "sub": tableau_user,
-        "scp": scopes
-        },
-        jwt_secret,
-        algorithm = "HS256",
-        headers = {
-        'kid': jwt_secret_id,
-        'iss': jwt_client_id
-        }
-    )
-
-    # authentication endpoint + request headers & payload
-    endpoint = f"{tableau_domain}/api/{tableau_api}/auth/signin"
-
-    headers = {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-    }
-
-    payload = {
-        "credentials": {
-        "jwt": token,
-        "site": {
-            "contentUrl": tableau_site,
-        }
-        }
-    }
-
-    response = requests.post(endpoint, headers=headers, json=payload)
-
-    # Check if the request was successful (status code 200)
-    if response.status_code == 200:
-        return response.json()
-    else:
-        error_message = (
-            f"Failed to authenticate to the Tableau site. "
-            f"Status code: {response.status_code}. Response: {response.text}"
-        )
-        raise RuntimeError(error_message)
-
-
 def prepare_prompt_inputs(data: dict, user_string: str) -> dict:
     """
     Prepare inputs for the prompt template with explicit, safe mapping.
@@ -265,49 +111,3 @@ def prepare_prompt_inputs(data: dict, user_string: str) -> dict:
         "data_table": data.get('data_table', ''),
         "user_input": user_string
     }
-
-
-def env_vars_simple_datasource_qa(
-    domain=None,
-    site=None,
-    jwt_client_id=None,
-    jwt_secret_id=None,
-    jwt_secret=None,
-    tableau_api_version=None,
-    tableau_user=None,
-    datasource_luid=None,
-    tooling_llm_model=None
-):
-    """
-    Retrieves Tableau configuration from environment variables if not provided as arguments.
-
-    Args:
-        domain (str, optional): Tableau domain
-        site (str, optional): Tableau site
-        jwt_client_id (str, optional): JWT client ID
-        jwt_secret_id (str, optional): JWT secret ID
-        jwt_secret (str, optional): JWT secret
-        tableau_api_version (str, optional): Tableau API version
-        tableau_user (str, optional): Tableau user
-        datasource_luid (str, optional): Datasource LUID
-        tooling_llm_model (str, optional): Tooling LLM model
-
-    Returns:
-        dict: A dictionary containing all the configuration values
-    """
-    # Load environment variables before accessing them
-    load_dotenv()
-
-    config = {
-        'domain': domain or os.environ.get('TABLEAU_DOMAIN'),
-        'site': site or os.environ.get('TABLEAU_SITE'),
-        'jwt_client_id': jwt_client_id or os.environ.get('TABLEAU_JWT_CLIENT_ID'),
-        'jwt_secret_id': jwt_secret_id or os.environ.get('TABLEAU_JWT_SECRET_ID'),
-        'jwt_secret': jwt_secret or os.environ.get('TABLEAU_JWT_SECRET'),
-        'tableau_api_version': tableau_api_version or os.environ.get('TABLEAU_API_VERSION'),
-        'tableau_user': tableau_user or os.environ.get('TABLEAU_USER'),
-        'datasource_luid': datasource_luid or os.environ.get('DATASOURCE_LUID'),
-        'tooling_llm_model': tooling_llm_model or os.environ.get('TOOLING_MODEL')
-    }
-
-    return config
